Change Lat Lon type to BigDecimal and include C3P0 to manage conx pool

+++ iBlue-Core/src/main/java/com/iblue/model/SpotInterface.java
@@ -1,9 +1,11 @@
 package com.iblue.model;
 
+import java.math.BigDecimal;
+
 public interface SpotInterface {
 
-	public float getLatitude(); // latitude in degrees
-	public float getLongitude(); // longitude in degrees
+	public BigDecimal getLatitude(); // latitude in degrees
+	public BigDecimal getLongitude(); // longitude in degrees
 	public String getMac();
 	public int getId();
 	public int getStatus();
